feat(rust): create 3 separate credential retriever types
