+++ implementations/rust/ockam/ockam_api/src/cli_state/storage/enrollments_repository_sql.rs
@@ -194,7 +194,7 @@ mod tests {
         // retrieve only the enrolled identities
         let result = repository.get_enrolled_identities().await?;
         assert_eq!(result.len(), 1);
-        assert_eq!(result[0].email(), Some(email));
+        assert_eq!(result[0].email(), &Some(email));
 
         // the first identity must be seen as enrolled
         let result = repository.is_identity_enrolled("identity1").await?;
