+++ implementations/rust/ockam/ockam_api/src/nodes/connection/project.rs
@@ -69,6 +69,7 @@ impl Instantiator for ProjectInstantiator {
                 tcp.route,
                 &self.identifier.clone(),
                 Some(vec![project_identifier]),
+                None,
                 self.timeout,
             )
             .await?;
