+++ implementations/rust/ockam/ockam_api/src/nodes/mod.rs
@@ -7,7 +7,7 @@ pub use service::background_node_client::*;
 pub use service::in_memory_node::*;
 pub use service::policy::*;
 /// The main node-manager service running on remote nodes
-pub use service::{IdentityOverride, NodeManager, NodeManagerWorker};
+pub use service::{NodeManager, NodeManagerWorker};
 
 /// A const address to bind and send messages to
 pub const NODEMANAGER_ADDR: &str = "_internal.nodemanager";
