+++ implementations/rust/ockam/ockam_api/src/nodes/service/in_memory_node.rs
@@ -145,7 +145,7 @@ impl InMemoryNode {
             .into_diagnostic()?;
 
         let trust_options = cli_state
-            .retrieve_trust_options(&project_name, &authority_identity, &authority_route, false)
+            .retrieve_trust_options(&project_name, &authority_identity, &authority_route, &None)
             .await
             .into_diagnostic()?;
