+++ implementations/rust/ockam/ockam_api/src/nodes/service/worker.rs
@@ -1,0 +1,293 @@
+use crate::nodes::models::policies::SetPolicyRequest;
+use crate::nodes::registry::KafkaServiceKind;
+use crate::nodes::service::{encode_response, TARGET};
+use crate::nodes::{InMemoryNode, NODEMANAGER_ADDR};
+use crate::DefaultAddress;
+use minicbor::Decoder;
+use ockam_core::api::{RequestHeader, Response};
+use ockam_core::{Address, Routed, Worker};
+use ockam_node::Context;
+use std::error::Error;
+use std::sync::Arc;
+
+#[derive(Clone)]
+pub struct NodeManagerWorker {
+    pub node_manager: Arc<InMemoryNode>,
+}
+
+impl NodeManagerWorker {
+    pub fn new(node_manager: Arc<InMemoryNode>) -> Self {
+        NodeManagerWorker { node_manager }
+    }
+
+    pub async fn stop(&self, ctx: &Context) -> ockam_core::Result<()> {
+        self.node_manager.stop(ctx).await?;
+        ctx.stop_worker(NODEMANAGER_ADDR).await?;
+        Ok(())
+    }
+}
+
+impl NodeManagerWorker {
+    //////// Request matching and response handling ////////
+
+    #[instrument(skip_all, fields(method = ?req.method(), path = req.path()))]
+    async fn handle_request(
+        &mut self,
+        ctx: &mut Context,
+        req: &RequestHeader,
+        dec: &mut Decoder<'_>,
+    ) -> ockam_core::Result<Vec<u8>> {
+        debug! {
+            target: TARGET,
+            id     = %req.id(),
+            method = ?req.method(),
+            path   = %req.path(),
+            body   = %req.has_body(),
+            "request"
+        }
+
+        use ockam_core::api::Method::*;
+        let path = req.path();
+        let path_segments = req.path_segments::<5>();
+        let method = match req.method() {
+            Some(m) => m,
+            None => todo!(),
+        };
+
+        let r = match (method, path_segments.as_slice()) {
+            // ==*== Basic node information ==*==
+            // TODO: create, delete, destroy remote nodes
+            (Get, ["node"]) => encode_response(req, self.get_node_status(ctx).await)?,
+
+            // ==*== Tcp Connection ==*==
+            (Get, ["node", "tcp", "connection"]) => self.get_tcp_connections(req).await.to_vec()?,
+            (Get, ["node", "tcp", "connection", address]) => {
+                encode_response(req, self.get_tcp_connection(address.to_string()).await)?
+            }
+            (Post, ["node", "tcp", "connection"]) => {
+                encode_response(req, self.create_tcp_connection(ctx, dec.decode()?).await)?
+            }
+            (Delete, ["node", "tcp", "connection"]) => {
+                encode_response(req, self.delete_tcp_connection(dec.decode()?).await)?
+            }
+
+            // ==*== Tcp Listeners ==*==
+            (Get, ["node", "tcp", "listener"]) => self.get_tcp_listeners(req).await.to_vec()?,
+            (Get, ["node", "tcp", "listener", address]) => {
+                encode_response(req, self.get_tcp_listener(address.to_string()).await)?
+            }
+            (Post, ["node", "tcp", "listener"]) => {
+                encode_response(req, self.create_tcp_listener(dec.decode()?).await)?
+            }
+            (Delete, ["node", "tcp", "listener"]) => {
+                encode_response(req, self.delete_tcp_listener(dec.decode()?).await)?
+            }
+
+            // ==*== Secure channels ==*==
+            (Get, ["node", "secure_channel"]) => {
+                encode_response(req, self.list_secure_channels().await)?
+            }
+            (Get, ["node", "secure_channel_listener"]) => {
+                encode_response(req, self.list_secure_channel_listener().await)?
+            }
+            (Post, ["node", "secure_channel"]) => {
+                encode_response(req, self.create_secure_channel(dec.decode()?, ctx).await)?
+            }
+            (Delete, ["node", "secure_channel"]) => {
+                encode_response(req, self.delete_secure_channel(dec.decode()?, ctx).await)?
+            }
+            (Get, ["node", "show_secure_channel"]) => {
+                encode_response(req, self.show_secure_channel(dec.decode()?).await)?
+            }
+            (Post, ["node", "secure_channel_listener"]) => encode_response(
+                req,
+                self.create_secure_channel_listener(dec.decode()?, ctx)
+                    .await,
+            )?,
+            (Delete, ["node", "secure_channel_listener"]) => encode_response(
+                req,
+                self.delete_secure_channel_listener(dec.decode()?, ctx)
+                    .await,
+            )?,
+            (Get, ["node", "show_secure_channel_listener"]) => {
+                encode_response(req, self.show_secure_channel_listener(dec.decode()?).await)?
+            }
+
+            // ==*== Services ==*==
+            (Post, ["node", "services", DefaultAddress::UPPERCASE_SERVICE]) => {
+                encode_response(req, self.start_uppercase_service(ctx, dec.decode()?).await)?
+            }
+            (Post, ["node", "services", DefaultAddress::ECHO_SERVICE]) => {
+                encode_response(req, self.start_echoer_service(ctx, dec.decode()?).await)?
+            }
+            (Post, ["node", "services", DefaultAddress::HOP_SERVICE]) => {
+                encode_response(req, self.start_hop_service(ctx, dec.decode()?).await)?
+            }
+            (Post, ["node", "services", DefaultAddress::KAFKA_OUTLET]) => encode_response(
+                req,
+                self.start_kafka_outlet_service(ctx, dec.decode()?).await,
+            )?,
+            (Delete, ["node", "services", DefaultAddress::KAFKA_OUTLET]) => encode_response(
+                req,
+                self.delete_kafka_service(ctx, dec.decode()?, KafkaServiceKind::Outlet)
+                    .await,
+            )?,
+            (Post, ["node", "services", DefaultAddress::KAFKA_CONSUMER]) => encode_response(
+                req,
+                self.start_kafka_consumer_service(ctx, dec.decode()?).await,
+            )?,
+            (Delete, ["node", "services", DefaultAddress::KAFKA_CONSUMER]) => encode_response(
+                req,
+                self.delete_kafka_service(ctx, dec.decode()?, KafkaServiceKind::Consumer)
+                    .await,
+            )?,
+            (Post, ["node", "services", DefaultAddress::KAFKA_PRODUCER]) => encode_response(
+                req,
+                self.start_kafka_producer_service(ctx, dec.decode()?).await,
+            )?,
+            (Delete, ["node", "services", DefaultAddress::KAFKA_PRODUCER]) => encode_response(
+                req,
+                self.delete_kafka_service(ctx, dec.decode()?, KafkaServiceKind::Producer)
+                    .await,
+            )?,
+            (Post, ["node", "services", DefaultAddress::KAFKA_DIRECT]) => encode_response(
+                req,
+                self.start_kafka_direct_service(ctx, dec.decode()?).await,
+            )?,
+            (Delete, ["node", "services", DefaultAddress::KAFKA_DIRECT]) => encode_response(
+                req,
+                self.delete_kafka_service(ctx, dec.decode()?, KafkaServiceKind::Direct)
+                    .await,
+            )?,
+            (Get, ["node", "services"]) => encode_response(req, self.list_services().await)?,
+            (Get, ["node", "services", service_type]) => {
+                encode_response(req, self.list_services_of_type(service_type).await)?
+            }
+
+            // ==*== Relay commands ==*==
+            (Get, ["node", "relay", alias]) => {
+                encode_response(req, self.show_relay(req, alias).await)?
+            }
+            (Get, ["node", "relay"]) => encode_response(req, self.get_relays(req).await)?,
+            (Delete, ["node", "relay", alias]) => {
+                encode_response(req, self.delete_relay(req, alias).await)?
+            }
+            (Post, ["node", "relay"]) => {
+                encode_response(req, self.create_relay(ctx, req, dec.decode()?).await)?
+            }
+
+            // ==*== Inlets & Outlets ==*==
+            (Get, ["node", "inlet"]) => encode_response(req, self.get_inlets().await)?,
+            (Get, ["node", "inlet", alias]) => encode_response(req, self.show_inlet(alias).await)?,
+            (Get, ["node", "outlet"]) => self.get_outlets(req).await.to_vec()?,
+            (Get, ["node", "outlet", addr]) => {
+                let addr: Address = addr.to_string().into();
+                encode_response(req, self.show_outlet(&addr).await)?
+            }
+            (Post, ["node", "inlet"]) => {
+                encode_response(req, self.create_inlet(ctx, dec.decode()?).await)?
+            }
+            (Post, ["node", "outlet"]) => {
+                encode_response(req, self.create_outlet(ctx, dec.decode()?).await)?
+            }
+            (Delete, ["node", "outlet", addr]) => {
+                let addr: Address = addr.to_string().into();
+                encode_response(req, self.delete_outlet(&addr).await)?
+            }
+            (Delete, ["node", "inlet", alias]) => {
+                encode_response(req, self.delete_inlet(alias).await)?
+            }
+            (Delete, ["node", "portal"]) => todo!(),
+
+            // ==*== Flow Controls ==*==
+            (Post, ["node", "flow_controls", "add_consumer"]) => {
+                encode_response(req, self.add_consumer(ctx, dec.decode()?).await)?
+            }
+
+            // ==*== Workers ==*==
+            (Get, ["node", "workers"]) => encode_response(req, self.list_workers(ctx).await)?,
+
+            // ==*== Policies ==*==
+            (Post, ["policy", action]) => {
+                let payload: SetPolicyRequest = dec.decode()?;
+                encode_response(
+                    req,
+                    self.add_policy(action, payload.resource, payload.expression)
+                        .await,
+                )?
+            }
+            (Get, ["policy", action]) => {
+                encode_response(req, self.get_policy(action, dec.decode()?).await)?
+            }
+            (Get, ["policy"]) => encode_response(req, self.list_policies(dec.decode()?).await)?,
+            (Delete, ["policy", action]) => {
+                encode_response(req, self.delete_policy(action, dec.decode()?).await)?
+            }
+
+            // ==*== Messages ==*==
+            (Post, ["v0", "message"]) => {
+                encode_response(req, self.send_message(ctx, dec.decode()?).await)?
+            }
+
+            // ==*== Catch-all for Unimplemented APIs ==*==
+            _ => {
+                warn!(%method, %path, "Called invalid endpoint");
+                Response::bad_request(req, &format!("Invalid endpoint: {} {}", method, path))
+                    .to_vec()?
+            }
+        };
+        Ok(r)
+    }
+}
+
+#[ockam::worker]
+impl Worker for NodeManagerWorker {
+    type Message = Vec<u8>;
+    type Context = Context;
+
+    async fn shutdown(&mut self, ctx: &mut Self::Context) -> ockam_core::Result<()> {
+        self.node_manager.medic_handle.stop_medic(ctx).await
+    }
+
+    async fn handle_message(
+        &mut self,
+        ctx: &mut Context,
+        msg: Routed<Vec<u8>>,
+    ) -> ockam_core::Result<()> {
+        let return_route = msg.return_route();
+        let body = msg.into_body()?;
+        let mut dec = Decoder::new(&body);
+        let req: RequestHeader = match dec.decode() {
+            Ok(r) => r,
+            Err(e) => {
+                error!("Failed to decode request: {:?}", e);
+                return Ok(());
+            }
+        };
+
+        let r = match self.handle_request(ctx, &req, &mut dec).await {
+            Ok(r) => r,
+            Err(err) => {
+                error! {
+                    target: TARGET,
+                    re     = %req.id(),
+                    method = ?req.method(),
+                    path   = %req.path(),
+                    code   = %err.code(),
+                    cause  = ?err.source(),
+                    "failed to handle request"
+                }
+                Response::internal_error(&req, &format!("failed to handle request: {err} {req:?}"))
+                    .to_vec()?
+            }
+        };
+        debug! {
+            target: TARGET,
+            re     = %req.id(),
+            method = ?req.method(),
+            path   = %req.path(),
+            "responding"
+        }
+        ctx.send(return_route, r).await
+    }
+}
