+++ implementations/rust/ockam/ockam_api/tests/latency.rs
@@ -44,6 +44,7 @@ pub fn measure_message_latency_two_nodes() {
                     None,
                     None,
                     None,
+                    None,
                 )
                 .await
                 .unwrap();
